[LM-18]: Updated migration script changed duration from int to double.

+++ src/main/resources/db/migration/V2__Create_leaves_table.sql
@@ -1,7 +1,7 @@
 CREATE TABLE leaves (
     id BIGSERIAL PRIMARY KEY,
     date DATE NOT NULL,
-    duration INT NOT NULL,
+    duration DOUBLE PRECISION NOT NULL,
     description VARCHAR(255),
     half_day VARCHAR(255),
     user_id BIGINT NOT NULL,
